Add `perks/perk_test.ts` for `parseIsAvailable`
