+++ deno.jsonc
@@ -3,9 +3,9 @@
   "tasks": {
     "lint": "deno lint",
     "fmt": "deno fmt",
-    "lock": "deno cache --lock-write bot/deps.ts",
-    "udd": "deno run -Ar --allow-read=. --allow-write=. --allow-net https://deno.land/x/udd/main.ts bot/deps.ts && deno task lock",
-    "all": "deno task udd && deno task lock && deno task lint && deno task fmt",
+    "lock": "deno cache --lock-write dev_deps.ts bot/deps.ts",
+    "udd": "deno run -Ar --allow-read=. --allow-write=. --allow-net https://deno.land/x/udd/main.ts dev_deps.ts bot/deps.ts && deno task lock",
+    "all": "deno task udd && deno task lock && deno task lint && deno task fmt && deno test",
     "start": "deno run -A --unstable main.ts"
   }
 }
