+++ dev_deps.ts
@@ -1,0 +1,1 @@
+export { assertEquals } from "https://deno.land/std@0.190.0/testing/asserts.ts";
